Msf-Session raises exception with netcat shells
When executing a msf-session command that utilize netcat shells,
the commands will fail and raise TypeError exceptions.
This commit adds a workaround so that if this exception occurs,
then the msf-session command tries to execute the command again
using a raw-write and reads the output using a raw-read.

Another change is, that it is now possible to use "write: True" and
"read: True" together. The command will first write and then read.